Merged PR 4008660: [Tooling] Test Image Build Pipeline Setup + Containerization
Added integration for containerization in the root project. This would form the basis of `One Click` build a test image for Kudu Pipeline.

This unifies the containerization and the KuduLite project

`docker build .` on the root directory would now directly produce a Docker Image for the Private Branch.

I have created a new ACR `appsvctest` : I'm evaluating about having the credentials for this ACR on workers as env variables so a test image produced by the pipeline can be consumed on the test stamp.