Merged PR 4019072: [WebSSH] Remove PM2 Dependency from webssh and move to per instance logs
This PR removes PM2 dependency for webssh. Instead it introduces a simple watcher script which runs as a daemon and detects if the webssh process is running in the background. If this script detects webssh process is not running it starts it.

This PR also removes the race condition where webssh error logs were recorded in a single file. It introduces per instance logs files.

This PR also ensures `stdout` logs go to `/dev/null`

Caveat: PM2 has a `flush` capability where we could flush the log files.(though I'm not sure if it worked as we have seen some cases where customers had 1Gb+ webssh stdout log files). **Here, we would have to introduce a Old Log Cleanup utility.** I have added a WI to record this: #9193781

Local Testing:
1. Killed the node process multiple times manually, and webssh kept running.
2. Added a few `console.error` logs in webssh app to check error logs were being recorded correctly
3. Tried changing the stdout logs location: just to test if the otput redicts were correct.

OGF Testing:
1. OGF has 30-40 apps where we test for `webssh` availaibility. I would be watching the results after the next upgrade for this. PS this is targeted for ANT 93

BONUS: pm2 was consuming ~30 Mb of Memory which was unnecessary. This should help us there.